backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.34.0
pymongo==4.16.0
redis==5.2.1
pyyaml==6.0.2
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.28.1
websockets==14.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
numpy>=1.24.0
sentence-transformers>=2.2.0
